Removed hGraphQL from edit co insured modules
