+++ Projects/App/Sources/Service/OctopusClientsImplementation/EditCoInsuredClientOctopus.swift
@@ -1,27 +1,8 @@
+import EditCoInsured
 import EditCoInsuredShared
 import Foundation
 import hCore
 import hGraphQL
-
-@MainActor
-public class EditCoInsuredService {
-    @Inject var service: EditCoInsuredClient
-
-    func sendMidtermChangeIntentCommit(commitId: String) async throws {
-        log.info("EditCoInsuredService: sendMidtermChangeIntentCommit", error: nil, attributes: nil)
-        return try await service.sendMidtermChangeIntentCommit(commitId: commitId)
-    }
-
-    func getPersonalInformation(SSN: String) async throws -> PersonalData? {
-        log.info("EditCoInsuredService: getPersonalInformation", error: nil, attributes: nil)
-        return try await service.getPersonalInformation(SSN: SSN)
-    }
-
-    func sendIntent(contractId: String, coInsured: [CoInsuredModel]) async throws -> Intent {
-        log.info("EditCoInsuredService: sendIntent", error: nil, attributes: nil)
-        return try await service.sendIntent(contractId: contractId, coInsured: coInsured)
-    }
-}
 
 public class EditCoInsuredClientOctopus: EditCoInsuredClient {
     @Inject var octopus: hOctopus
