+++ Projects/App/Sources/Service/OctopusClientsImplementation/EditCoInsuredSharedClientOctopus.swift
@@ -1,16 +1,7 @@
+import EditCoInsuredShared
 import Foundation
 import hCore
 import hGraphQL
-
-@MainActor
-public class EditCoInsuredSharedService {
-    @Inject var service: EditCoInsuredSharedClient
-
-    public func fetchContracts() async throws -> [Contract] {
-        log.info("EditCoInsuredSharedService: fetchContracts", error: nil, attributes: nil)
-        return try await service.fetchContracts()
-    }
-}
 
 public class EditCoInsuredSharedClientOctopus: EditCoInsuredSharedClient {
     @Inject var octopus: hOctopus
