+++ Projects/EditCoInsured/Sources/Models/IntentData.swift
@@ -7,4 +7,18 @@ public struct Intent: Sendable {
     let newPremium: MonetaryAmount
     let id: String
     let state: String
+
+    public init(
+        activationDate: String,
+        currentPremium: MonetaryAmount,
+        newPremium: MonetaryAmount,
+        id: String,
+        state: String
+    ) {
+        self.activationDate = activationDate
+        self.currentPremium = currentPremium
+        self.newPremium = newPremium
+        self.id = id
+        self.state = state
+    }
 }
