+++ Projects/EditCoInsured/Sources/Service/DemoImplementation/EditCoInsuredClientDemo.swift
@@ -1,7 +1,6 @@
 import EditCoInsuredShared
 import Foundation
 import hCore
-import hGraphQL
 
 public class EditCoInsuredClientDemo: EditCoInsuredClient {
     public func sendMidtermChangeIntentCommit(commitId: String) async throws {
